src/api/requirements.txt
# API requirements
fastapi>=0.115.12
uvicorn>=0.34.2
python-dotenv>=1.1.0
supabase>=2.15.1
pydantic>=2.11.4
python-jose>=3.4.0
passlib>=1.7.4
python-multipart>=0.0.20
httpx>=0.28.1
bcrypt>=4.3.0
email-validator>=2.2.0

# Testing requirements
pytest>=8.3.5
pytest-asyncio>=0.26.0
pytest-cov>=6.1.1

